+++ src/controllers/UsersController.php
@@ -602,7 +602,13 @@ class UsersController extends BaseController
 
 					if (craft()->userSession->isAdmin())
 					{
-						$statusActions[] = array('id' => 'copy-passwordreset-url', 'label' => Craft::t('Copy activation URL'));
+						// If they already have a password (like from a front-end user registration), no
+						// need to show the Copy Password Reset URL option
+						if (!$variables['account']->password)
+						{
+							$statusActions[] = ['id' => 'copy-passwordreset-url', 'label' => Craft::t('Copy activation URL')];
+						}
+
 						$statusActions[] = array('action' => 'users/activateUser', 'label' => Craft::t('Activate account'));
 					}